Move base.formatYAxis to intl.formatNumber
